+++ blog-vue/src/App.vue
@@ -1,7 +1,7 @@
 <template>
   <div id="nav">
     <router-link to="/">Home</router-link> |
-    <router-link to="/about">About</router-link>
+    <router-link to="/Blogs">Blogs</router-link> | <!--Or name it - All posts -->
   </div>
   <router-view/>
 </template>
