+++ blog-vue/src/views/Article.vue
@@ -1,0 +1,17 @@
+<template>
+    <!-- article title-->
+    <div>
+        
+    </div>
+</template>
+
+<script>
+    export default {
+        name: 'Article'
+    }
+
+</script>
+
+<style>
+
+</style>
