+++ blog-vue/src/views/Header.vue
@@ -1,0 +1,17 @@
+<template>
+<header>
+    <img id="logo" alt="Student logo" src="../assets/student.png">
+    </header>
+</template>
+
+<script>
+    export default {
+        name: 'Header',
+        
+    }
+
+</script>
+
+<style>
+
+</style>
