--- blog-vue/src/views/myPart.txt
@@ -1,1 +1,0 @@
-comit
